:bug: fix(core): stop auto-zoom on input fields on iOS browsers (#223)
Since Safari/iOS mobile browsers automatically zoom on input fields when the
font size is under 16px, since we have a responsive design, we can remove this
by setting the maximum-scale of the viewport.

Notably, this may be not be great for accessibility, so might need changing in
the future.